Move HTTP view registration to async_setup in HTML5 integration
