+++ homeassistant/components/html5/const.py
@@ -25,3 +25,5 @@ ATTR_TIMESTAMP = "timestamp"
 ATTR_TTL = "ttl"
 ATTR_URGENCY = "urgency"
 ATTR_VIBRATE = "vibrate"
+ATTR_SUBSCRIPTION = "subscription"
+ATTR_ENDPOINT = "endpoint"
