+++ tests/components/html5/conftest.py
@@ -55,9 +55,11 @@ def mock_config_entry() -> MockConfigEntry:
 def mock_load_config() -> Generator[MagicMock]:
     """Mock load config."""
 
-    with patch(
-        "homeassistant.components.html5.notify._load_config", return_value={}
-    ) as mock_load_config:
+    with (
+        patch(
+            "homeassistant.components.html5.notify._load_config", return_value={}
+        ) as mock_load_config,
+    ):
         yield mock_load_config
 
 
@@ -91,6 +93,7 @@ def mock_jwt() -> Generator[MagicMock]:
 
     with (
         patch("homeassistant.components.html5.notify.jwt") as mock_client,
+        patch("homeassistant.components.html5.http.jwt", new=mock_client),
     ):
         mock_client.encode.return_value = "JWT"
         mock_client.decode.return_value = {"target": "device"}
@@ -122,4 +125,15 @@ def mock_vapid() -> Generator[MagicMock]:
             "urgency": "normal",
             "priority": "normal",
         }
+        yield mock_client
+
+
+@pytest.fixture
+def mock_save() -> Generator[MagicMock]:
+    """Mock save_json."""
+
+    with (
+        patch("homeassistant.components.html5.http.save_json") as mock_client,
+        patch("homeassistant.components.html5.notify.save_json", new=mock_client),
+    ):
         yield mock_client
