+++ tests/components/html5/test_notify.py
@@ -31,7 +31,8 @@ from homeassistant.components.html5.const import (
     ATTR_VIBRATE,
     SERVICE_DISMISS,
 )
-from homeassistant.components.html5.notify import ATTR_ACTION, ATTR_DISMISS, DEFAULT_TTL
+from homeassistant.components.html5.http import ATTR_ACTION
+from homeassistant.components.html5.notify import ATTR_DISMISS, DEFAULT_TTL
 from homeassistant.components.html5.services import SERVICE_DISMISS_MESSAGE
 from homeassistant.components.notify import (
     ATTR_DATA,
@@ -347,6 +348,7 @@ async def test_registering_new_device_view(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML view works."""
     await async_setup_component(hass, "http", {})
@@ -359,8 +361,7 @@ async def test_registering_new_device_view(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
 
     assert resp.status == HTTPStatus.OK
     assert len(mock_save.mock_calls) == 1
@@ -372,6 +373,7 @@ async def test_registering_new_device_view_with_name(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML view works with name attribute."""
     await async_setup_component(hass, "http", {})
@@ -387,8 +389,7 @@ async def test_registering_new_device_view_with_name(
     SUB_WITH_NAME = SUBSCRIPTION_1.copy()
     SUB_WITH_NAME["name"] = "test device"
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUB_WITH_NAME))
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUB_WITH_NAME))
 
     assert resp.status == HTTPStatus.OK
     assert len(mock_save.mock_calls) == 1
@@ -400,6 +401,7 @@ async def test_registering_new_device_expiration_view(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML view works."""
     await async_setup_component(hass, "http", {})
@@ -412,8 +414,7 @@ async def test_registering_new_device_expiration_view(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
 
     assert resp.status == HTTPStatus.OK
     assert mock_save.mock_calls[0][1][1] == {"unnamed device": SUBSCRIPTION_4}
@@ -424,6 +425,7 @@ async def test_registering_new_device_fails_view(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test subs. are not altered when registering a new device fails."""
     await async_setup_component(hass, "http", {})
@@ -435,11 +437,9 @@ async def test_registering_new_device_fails_view(
     assert config_entry.state is ConfigEntryState.LOADED
 
     client = await hass_client()
-    with patch(
-        "homeassistant.components.html5.notify.save_json",
-        side_effect=HomeAssistantError(),
-    ):
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
+    mock_save.side_effect = (HomeAssistantError(),)
+
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
 
     assert resp.status == HTTPStatus.INTERNAL_SERVER_ERROR
 
@@ -449,6 +449,7 @@ async def test_registering_existing_device_view(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test subscription is updated when registering existing device."""
     await async_setup_component(hass, "http", {})
@@ -461,9 +462,8 @@ async def test_registering_existing_device_view(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
+    await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
 
     assert resp.status == HTTPStatus.OK
     mock_save.assert_called_with(
@@ -476,6 +476,7 @@ async def test_registering_existing_device_view_with_name(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test subscription is updated when reg'ing existing device with name."""
     await async_setup_component(hass, "http", {})
@@ -491,9 +492,8 @@ async def test_registering_existing_device_view_with_name(
     SUB_WITH_NAME = SUBSCRIPTION_1.copy()
     SUB_WITH_NAME["name"] = "test device"
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        await client.post(REGISTER_URL, data=json.dumps(SUB_WITH_NAME))
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
+    await client.post(REGISTER_URL, data=json.dumps(SUB_WITH_NAME))
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
 
     assert resp.status == HTTPStatus.OK
 
@@ -507,6 +507,7 @@ async def test_registering_existing_device_fails_view(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test sub. is not updated when registering existing device fails."""
     await async_setup_component(hass, "http", {})
@@ -519,10 +520,9 @@ async def test_registering_existing_device_fails_view(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
-        mock_save.side_effect = HomeAssistantError
-        resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
+    await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_1))
+    mock_save.side_effect = HomeAssistantError
+    resp = await client.post(REGISTER_URL, data=json.dumps(SUBSCRIPTION_4))
 
     assert resp.status == HTTPStatus.INTERNAL_SERVER_ERROR
 
@@ -532,6 +532,7 @@ async def test_registering_new_device_validation(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test various errors when registering a new device."""
     await async_setup_component(hass, "http", {})
@@ -553,11 +554,11 @@ async def test_registering_new_device_validation(
     resp = await client.post(REGISTER_URL, data=json.dumps({"browser": "chrome"}))
     assert resp.status == HTTPStatus.BAD_REQUEST
 
-    with patch("homeassistant.components.html5.notify.save_json", return_value=False):
-        resp = await client.post(
-            REGISTER_URL,
-            data=json.dumps({"browser": "chrome", "subscription": "sub info"}),
-        )
+    mock_save.return_value = False
+    resp = await client.post(
+        REGISTER_URL,
+        data=json.dumps({"browser": "chrome", "subscription": "sub info"}),
+    )
     assert resp.status == HTTPStatus.BAD_REQUEST
 
 
@@ -566,6 +567,7 @@ async def test_unregistering_device_view(
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
     load_config: MagicMock,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML unregister view works."""
     load_config.return_value = {
@@ -582,11 +584,10 @@ async def test_unregistering_device_view(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        resp = await client.delete(
-            REGISTER_URL,
-            data=json.dumps({"subscription": SUBSCRIPTION_1["subscription"]}),
-        )
+    resp = await client.delete(
+        REGISTER_URL,
+        data=json.dumps({"subscription": SUBSCRIPTION_1["subscription"]}),
+    )
 
     assert resp.status == HTTPStatus.OK
     assert len(mock_save.mock_calls) == 1
@@ -600,6 +601,7 @@ async def test_unregister_device_view_handle_unknown_subscription(
     hass: HomeAssistant,
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML unregister view handles unknown subscriptions."""
     await async_setup_component(hass, "http", {})
@@ -612,11 +614,10 @@ async def test_unregister_device_view_handle_unknown_subscription(
 
     client = await hass_client()
 
-    with patch("homeassistant.components.html5.notify.save_json") as mock_save:
-        resp = await client.delete(
-            REGISTER_URL,
-            data=json.dumps({"subscription": SUBSCRIPTION_3["subscription"]}),
-        )
+    resp = await client.delete(
+        REGISTER_URL,
+        data=json.dumps({"subscription": SUBSCRIPTION_3["subscription"]}),
+    )
 
     assert resp.status == HTTPStatus.OK, resp.response
     assert len(mock_save.mock_calls) == 0
@@ -627,6 +628,7 @@ async def test_unregistering_device_view_handles_save_error(
     hass_client: ClientSessionGenerator,
     config_entry: MockConfigEntry,
     load_config: MagicMock,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the HTML unregister view handles save errors."""
     load_config.return_value = {
@@ -643,14 +645,12 @@ async def test_unregistering_device_view_handles_save_error(
 
     client = await hass_client()
 
-    with patch(
-        "homeassistant.components.html5.notify.save_json",
-        side_effect=HomeAssistantError(),
-    ):
-        resp = await client.delete(
-            REGISTER_URL,
-            data=json.dumps({"subscription": SUBSCRIPTION_1["subscription"]}),
-        )
+    mock_save.side_effect = HomeAssistantError()
+
+    resp = await client.delete(
+        REGISTER_URL,
+        data=json.dumps({"subscription": SUBSCRIPTION_1["subscription"]}),
+    )
 
     assert resp.status == HTTPStatus.INTERNAL_SERVER_ERROR, resp.response
 
@@ -785,6 +785,7 @@ async def test_send_fcm_expired(
     config_entry: MockConfigEntry,
     load_config: MagicMock,
     mock_wp: AsyncMock,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the FCM target is removed when expired."""
     load_config.return_value = {"device": SUBSCRIPTION_5}
@@ -796,15 +797,13 @@ async def test_send_fcm_expired(
 
     assert config_entry.state is ConfigEntryState.LOADED
     mock_wp.send_async.return_value.status = 410
-    with (
-        patch("homeassistant.components.html5.notify.save_json") as mock_save,
-    ):
-        await hass.services.async_call(
-            "notify",
-            "html5",
-            {"message": "Hello", "target": ["device"], "data": {"icon": "beer.png"}},
-            blocking=True,
-        )
+
+    await hass.services.async_call(
+        "notify",
+        "html5",
+        {"message": "Hello", "target": ["device"], "data": {"icon": "beer.png"}},
+        blocking=True,
+    )
     # "device" should be removed when expired.
     mock_save.assert_called_once_with(hass.config.path(html5.REGISTRATIONS_FILE), {})
 
@@ -817,6 +816,7 @@ async def test_send_fcm_expired_save_fails(
     load_config: MagicMock,
     caplog: pytest.LogCaptureFixture,
     mock_wp: AsyncMock,
+    mock_save: MagicMock,
 ) -> None:
     """Test that the FCM target remains after expiry if save_json fails."""
     load_config.return_value = {"device": SUBSCRIPTION_5}
@@ -828,18 +828,13 @@ async def test_send_fcm_expired_save_fails(
 
     assert config_entry.state is ConfigEntryState.LOADED
     mock_wp.send_async.return_value.status = 410
-    with (
-        patch(
-            "homeassistant.components.html5.notify.save_json",
-            side_effect=HomeAssistantError(),
-        ),
-    ):
-        await hass.services.async_call(
-            "notify",
-            "html5",
-            {"message": "Hello", "target": ["device"], "data": {"icon": "beer.png"}},
-            blocking=True,
-        )
+    mock_save.side_effect = HomeAssistantError
+    await hass.services.async_call(
+        "notify",
+        "html5",
+        {"message": "Hello", "target": ["device"], "data": {"icon": "beer.png"}},
+        blocking=True,
+    )
     # "device" should still exist if save fails.
     assert "Error saving registration" in caplog.text
 
@@ -974,6 +969,7 @@ async def test_send_message_save_fails(
     webpush_async: AsyncMock,
     load_config: MagicMock,
     caplog: pytest.LogCaptureFixture,
+    mock_save: MagicMock,
 ) -> None:
     """Test sending a message with channel expired but saving registration fails."""
     load_config.return_value = {"my-desktop": SUBSCRIPTION_1}
@@ -987,13 +983,8 @@ async def test_send_message_save_fails(
     webpush_async.side_effect = (
         WebPushException("", response=Mock(status=HTTPStatus.GONE)),
     )
-    with (
-        patch(
-            "homeassistant.components.html5.notify.save_json",
-            side_effect=HomeAssistantError,
-        ),
-        pytest.raises(HomeAssistantError) as e,
-    ):
+    mock_save.side_effect = HomeAssistantError
+    with pytest.raises(HomeAssistantError) as e:
         await hass.services.async_call(
             NOTIFY_DOMAIN,
             SERVICE_SEND_MESSAGE,
